% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.SelInd}
\alias{as.data.frame.SelInd}
\title{Function to coerce the results of a SelInd object to a data.frame.}
\usage{
\method{as.data.frame}{SelInd}(x)
}
\arguments{
\item{x}{}
}
\value{
a data.frame with traits in rows and result vectors in columns
}
\description{
Function to coerce the results of a SelInd object to a data.frame.
}
