% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SelInd}
\alias{print.SelInd}
\title{Function to nicely print a SelInd object}
\usage{
\method{print}{SelInd}(x, ...)
}
\arguments{
\item{x}{An object of class SelInd}

\item{...}{does nothing, only for compatibility with the generic function}
}
\description{
Function to nicely print a SelInd object
}
\examples{
tn <- c("RZM", "RZN", "RZEo")
G <- matrix(
    c(1.0,0.13,0.13,
    0.13,1.0,0.23,
    0.13,0.23,1.0),
    3, 3, dimnames = list(tn,tn)
    ) * 144
w <- c(0.7, 0.3, 0)
names(w) <- tn
r2 <- c(0.743, 0.673)
names(r2) <- tn[1:2]
res <- SelInd(
  w = w,
  G = G,
  r2 = r2
)
print(res)

}
