% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desiredGain.R
\name{SelInd}
\alias{SelInd}
\title{Function to calculate selection index}
\usage{
SelInd(w, G, r, H = NULL, i = NULL, h2 = NULL, d_real = NULL)
}
\arguments{
\item{w}{Numeric vector of n economic weights. Traits present in G, but not part of the index need to be coded as 0. If traits of G are missing, they will be added automatically with zero weight. Required.}

\item{G}{Named n*n genetic variance- covariance matrix. Dimnames of G need to match E/r to ensure correct sorting.  Required.}

\item{r}{Named numeric vector of reliabilites with length m. Required.}

\item{H}{Named m*m matrix of empiric correlations between estimated breeding values.}

\item{i}{Selection intensity}

\item{h2}{named numeric vector of length n containing heritabilities for the traits}

\item{d_real}{named numeric vector of length n containing the observed composition of the genetic gain scaled in genetic standard deviations. If sum(d_real) != 1, it will be rescaled.}
}
\value{
A list of class SelInd
}
\description{
Function to calculate selection index
}
\details{
The framework allows to have less traits in the selection index than in the breeding goal (m < n). Calculation of realized economic weights, however, requires m == n.
}
\examples{

SelInd(
w = c(A=0.2,B=1,C=0,D=3),
G = matrix(c(1,0,0,0,
             0,2,0,0,
             0,0,1,0.5,
             0,0,0.5,1),4,4,
           dimnames = list(c('A','C','B','D'),c('A','C','B','D'))),
r = c(D=0.2,A=0.7),
i = 0.02,
h2 = c(A=0.2,B=0.2,C=0.2,D=0.3)
)

}
