% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SelInd}
\alias{print.SelInd}
\title{Function to nicely print a SelInd object}
\usage{
\method{print}{SelInd}(x)
}
\arguments{
\item{x}{An object of class SelInd}
}
\description{
Function to nicely print a SelInd object
}
