% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.SelInd}
\alias{summary.SelInd}
\title{Function to summarize the content of a selInd object.}
\usage{
\method{summary}{SelInd}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Function to summarize the content of a selInd object.
}
