% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.SelInd}
\alias{summary.SelInd}
\title{Function to summarize the content of a SelInd object.}
\usage{
\method{summary}{SelInd}(object, ...)
}
\arguments{
\item{object}{An object of class SelInd}

\item{...}{does nothing, only for compatibility with the generic function}
}
\description{
Function to summarize the content of a SelInd object.
}
